vivace_res/requirements.txt
Django
psycopg2
Pillow
django-admin-rangefilter
django-csp